+++ src/views/deveops/app/index.vue
@@ -1,0 +1,16 @@
+<template>
+    <section class="app-main" style="min-height: 100%">
+      <left-sidebar class="sidebar-container"></left-sidebar>
+      <router-view></router-view>
+    </section>
+</template>
+
+<script>
+  import LeftSidebar from '@/views/deveops/sidebar/left/index'
+  export default {
+    name: 'AppMain',
+    components: {
+      LeftSidebar
+    },
+  }
+</script>
