+++ src/views/deveops/console/enumerate.vue
