+++ src/views/deveops/console/terminal.vue
